- renderComments и initAddCommentListener теперь без аргументов - Переименована initEventListeners → initAddCommentListener - Разделены обработчики на 3 функции: добавление, лайки, цитирование - Удалена toggleLike из utils.js, код перенесен в обработчик лайков - Обновлены импорты между модулями"
