+++ main.js
@@ -1,7 +1,7 @@
 import { comments } from './modules/data.js';
 import { getComments } from './modules/api.js';
 import { renderComments } from './modules/render.js';
-import { initEventListeners } from './modules/listeners.js';
+import { initAddCommentListener } from './modules/listeners.js'; // импорт изменен
 
 const commentsList = document.querySelector('.comments');
 const commentsLoading = document.getElementById('comments-loading');
@@ -25,7 +25,7 @@ async function loadComments() {
         comments.length = 0;
         const freshComments = await getComments();
         comments.push(...freshComments);
-        renderComments(comments, commentsList);
+        renderComments(); // без аргументов!
         
     } catch (error) {
         commentsList.innerHTML = '<div class="error">Не удалось загрузить комментарии</div>';
@@ -36,4 +36,6 @@ async function loadComments() {
 }
 
 loadComments();
-initEventListeners(commentsList, loadComments);
+initAddCommentListener(); // вызов изменен и без аргументов
+
+export { loadComments };
