+++ modules/listeners.js
@@ -1,9 +1,10 @@
+import { loadComments } from '../main.js'; // импортируем LoadComments
 import { comments, replyingTo } from './data.js';
 import { addComment } from './api.js';
 import { renderComments } from './render.js';
 import { escapeHtml } from './utils.js';
 
-export function initEventListeners(commentsList, loadComments) {
+export function initAddCommentListener() {
     const nameInput = document.querySelector('.add-form-name');
     const textInput = document.querySelector('.add-form-text');
     const addButton = document.querySelector('.add-form-button');
@@ -27,7 +28,7 @@ export function initEventListeners(commentsList, loadComments) {
             textInput.value = '';
             nameInput.classList.remove('error');
             textInput.classList.remove('error');
-            await loadComments();
+            await loadComments(); // используем импортированную функцию
         } catch (error) {
             alert(error.message);
         } finally {
@@ -42,5 +43,41 @@ export function initEventListeners(commentsList, loadComments) {
 
     textInput.addEventListener('input', () => {
         if (textInput.value.trim()) textInput.classList.remove('error');
+    });
+}
+
+export function initLikeListeners() {
+    document.querySelectorAll('.like-button').forEach(button => {
+        button.addEventListener('click', (event) => {
+            const commentId = parseInt(event.target.closest('.comment').dataset.id);
+            const comment = comments.find(c => c.id === commentId);
+
+            if (comment) {
+                comment.isLiked = !comment.isLiked;
+                comment.likes += comment.isLiked ? 1 : -1;
+                renderComments();
+            }
+        });
+    });
+}
+
+export function initQuoteListeners() {
+    document.querySelectorAll('.comment').forEach(commentElement => {
+        commentElement.addEventListener('click', function (event) {
+            if (event.target.closest('.like-button')) {
+                return;
+            }
+
+            const commentId = parseInt(this.dataset.id);
+            const comment = comments.find(c => c.id === commentId);
+
+            if (comment) {
+                const textInput = document.querySelector('.add-form-text');
+                const quoteText = `> ${comment.author.name}: ${comment.text}\n\n`;
+                textInput.value = quoteText;
+                textInput.focus();
+                textInput.scrollIntoView({ behavior: 'smooth' });
+            }
+        });
     });
 }
