+++ modules/render.js
@@ -1,6 +1,9 @@
-import { toggleLike } from './utils.js';
+import { comments } from './data.js';
+import { initLikeListeners, initQuoteListeners } from './listeners.js';
 
-export function renderComments(comments, commentsList) {
+export function renderComments() {
+  const commentsList = document.querySelector('.comments');
+
   commentsList.innerHTML = comments.map(comment => `
     <li class="comment" data-id="${comment.id}">
       <div class="comment-header">
@@ -19,31 +22,7 @@ export function renderComments(comments, commentsList) {
     </li>
   `).join('');
 
-  document.querySelectorAll('.like-button').forEach(button => {
-    button.addEventListener('click', (event) => {
-      const commentId = parseInt(event.target.closest('.comment').dataset.id);
-      toggleLike(comments, commentId);
-      renderComments(comments, commentsList);
-    });
-  });
-
-
-  document.querySelectorAll('.comment').forEach(commentElement => {
-    commentElement.addEventListener('click', function (event) {
-      if (event.target.closest('.like-button')) {
-        return;
-      }
-
-      const commentId = parseInt(this.dataset.id);
-      const comment = comments.find(c => c.id === commentId);
-
-      if (comment) {
-        const textInput = document.querySelector('.add-form-text');
-        const quoteText = `> ${comment.author.name}: ${comment.text}\n\n`;
-        textInput.value = quoteText;
-        textInput.focus();
-        textInput.scrollIntoView({ behavior: 'smooth' });
-      }
-    });
-  });
+  // вызываем функции из listeners.js вместо кода здесь
+  initLikeListeners();
+  initQuoteListeners();
 }
