+++ modules/utils.js
@@ -6,12 +6,3 @@ export function escapeHtml(unsafe) {
         .replace(/"/g, "&quot;")
         .replace(/'/g, "&#039;");
 }
-
-export function toggleLike(comments, commentId) {
-    const comment = comments.find(c => c.id === commentId);
-    if (comment) {
-        comment.isLiked = !comment.isLiked;
-        comment.likes += comment.isLiked ? 1 : -1;
-    }
-    return comments;
-}
